cmake.sh/configure.sh => preprocess.sh

+++ apps/respack/install.sh
@@ -35,9 +35,11 @@ rm -rf $LOG
 cd ${BUILD_DIR}/${__NAME__}-${__VERSION__}
 start_info | tee -a $LOG
 
-echo "[cmake]" | tee -a $LOG
-rm -rf build && mkdir -p build && cd build
-check sh $CONFIG_DIR/cmake.sh
+echo "[preprocess]" | tee -a $LOG
+if [ -f CMakeLists.txt ]; then
+  rm -rf build && mkdir -p build && cd build
+fi
+check sh $CONFIG_DIR/preprocess.sh
 
 echo "[make]" | tee -a $LOG
 check make | tee -a $LOG || exit 1
